can add user working days to screen from the database

+++ src/main/java/net/webapp/Shift.java
@@ -1,39 +1,39 @@
 package net.webapp;
 
 public class Shift {
-    int id;
-    int userId;
-    int dayId;
+    Long id;
+    Long userId;
+    Long dayId;
 
     Shift(){}
 
-    public Shift(int id, int userId, int dayId) {
+    public Shift(Long id, Long userId, Long dayId) {
         this.id = id;
         this.userId = userId;
         this.dayId = dayId;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getDayId() {
+    public Long getDayId() {
         return dayId;
     }
 
-    public void setDayId(int dayId) {
+    public void setDayId(Long dayId) {
         this.dayId = dayId;
     }
 }
